KOGITO-4007 - separated ITs per PMML model
