+++ explainability/explainability-core/src/main/java/org/kie/kogito/explainability/utils/LocalSaliencyStability.java
@@ -20,6 +20,7 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
+import java.util.concurrent.ConcurrentHashMap;
 
 /**
  * Local {@code Saliency} stability evaluation result.
@@ -29,9 +30,9 @@ public class LocalSaliencyStability {
     private final Map<String, Map<Integer, SaliencyFrequencyMetadata>> map;
 
     public LocalSaliencyStability(Set<String> decisions) {
-        map = new HashMap<>();
+        map = new ConcurrentHashMap<>();
         for (String k : decisions) {
-            map.put(k, new HashMap<>());
+            map.put(k, new ConcurrentHashMap<>());
         }
     }
